+++ scrapper/src/main/java/edu/java/models/Chat.java
@@ -1,0 +1,4 @@
+package edu.java.models;
+
+public class Chat {
+}
