+++ scrapper/src/main/java/edu/java/models/GenericResponse.java
@@ -1,4 +1,4 @@
-package edu.java.dto;
+package edu.java.models;
 
 import edu.java.api.dto.response.ApiErrorResponse;
