+++ scrapper/src/main/java/edu/java/models/LinkData.java
@@ -1,4 +1,4 @@
-package edu.java.dto;
+package edu.java.models;
 
 import java.net.URL;
 import java.time.OffsetDateTime;
