+++ scrapper/src/main/java/edu/java/models/LinkUpdate.java
@@ -1,4 +1,4 @@
-package edu.java.dto;
+package edu.java.models;
 
 import java.net.URI;
 import java.util.List;
